# Stock-Dashboard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Warehouse Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List all materials
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: warehouseId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a new material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get a material by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
    patch:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /materials/{id}/stock-adjust:
    post:
      operationId: adjustMaterialStock
      tags: [materials]
      summary: Adjust stock (in/out/adjustment)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustInput"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
...
  /material-requests:
    get:
      operationId: listMaterialRequests
      tags: [material-requests]
      summary: List material requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: requesterId
          in: query
          schema:
            type: ["integer", "null"]
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of material requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaterialRequest"
    post:
      operationId: createMaterialRequest
      tags: [material-requests]
      summary: Create a new material request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialRequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialRequest"
  /material-requests/{id}:
    get:
      operationId: getMaterialRequest
      tags: [material-requests]
      summary: Get a material request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialRequest"
    delete:
      operationId: deleteMaterialRequest
      tags: [material-requests]
      summary: Delete a material request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /material-requests/{id}/approve:
    post:
      operationId: approveMaterialRequest
      tags: [material-requests]
      summary: Approve a material request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialRequest"
  /material-requests/{id}/reject:
    post:
      operationId: rejectMaterialRequest
      tags: [material-requests]
      summary: Reject a material request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialRequest"

  # ── TOOL REQUESTS ────────────────────────────────────────────────────────────
...
  /material-requests/{id}/release:
    post:
      operationId: releaseMaterialRequest
      tags: [material-requests]
      summary: Release/fulfil an approved material request (triggers stock-out)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseInput"
      responses:
        "200":
          description: Released request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialRequest"

  # ── REPORTS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Material:
      type: object
      required:
        - id
        - code
        - name
        - category
        - unit
        - currentStock
        - minimumStock
        - maximumStock
        - createdAt
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        category:
          type: string
        subCategory:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        specification:
          type: ["string", "null"]
        unit:
          type: string
        minimumStock:
          type: integer
        maximumStock:
          type: integer
        currentStock:
          type: integer
        reservedStock:
          type: integer
        rackLocation:
          type: ["string", "null"]
        warehouseZone:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        unitPrice:
          type: ["number", "null"]
        isLowStock:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    MaterialInput:
      type: object
      required: [name, category, unit, minimumStock, maximumStock]
      properties:
        name:
          type: string
        category:
          type: string
        subCategory:
          type: string
        brand:
          type: string
        specification:
          type: string
        unit:
          type: string
        minimumStock:
          type: integer
        maximumStock:
          type: integer
        currentStock:
          type: integer
        rackLocation:
          type: string
        warehouseZone:
          type: string
        supplier:
          type: string
        unitPrice:
          type: number
    MaterialUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        subCategory:
          type: string
        brand:
          type: string
        specification:
          type: string
        unit:
          type: string
        minimumStock:
          type: integer
        maximumStock:
          type: integer
        rackLocation:
          type: string
        warehouseZone:
          type: string
        supplier:
          type: string
        unitPrice:
          type: number
    StockAdjustInput:
      type: object
      required: [type, quantity, reason]
      properties:
        type:
          type: string
          enum: [stock_in, stock_out, adjustment]
        quantity:
          type: integer
        reason:
          type: string
        reference:
          type: string

    # ── Tool ───────────────────────────────────────────────────────────────────
...
    MaterialRequest:
      type: object
      required:
        - id
        - requestNumber
        - requesterName
        - department
        - status
        - priority
        - items
        - createdAt
      properties:
        id:
          type: integer
        requestNumber:
          type: string
        requesterName:
          type: string
        requesterId:
          type: ["integer", "null"]
        department:
          type: string
        workLocation:
          type: ["string", "null"]
        jobType:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [draft, pending, approved, rejected, released, completed]
        neededDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        approvedBy:
          type: ["string", "null"]
        approvalNotes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/RequestItem"
        createdAt:
          type: string
        updatedAt:
          type: string
    MaterialRequestInput:
      type: object
      required: [requesterName, department, priority, items]
      properties:
        requesterName:
          type: string
        requesterId:
          type: integer
        department:
          type: string
        workLocation:
          type: string
        jobType:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        neededDate:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RequestItemInput"
    RequestItem:
      type: object
      required: [id, materialId, materialName, materialCode, quantity, unit]
      properties:
        id:
          type: integer
        materialId:
          type: integer
        materialName:
          type: string
        materialCode:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        notes:
          type: ["string", "null"]
    RequestItemInput:
      type: object
      required: [materialId, quantity]
      properties:
        materialId:
          type: integer
        quantity:
          type: integer
        notes:
          type: string

    # ── Tool Request ───────────────────────────────────────────────────────────
...
    ApprovalInput:
      type: object
      required: [approvedBy]
      properties:
        approvedBy:
          type: string
        notes:
          type: string

    # ── Stock Movement ─────────────────────────────────────────────────────────
...
    ReleaseInput:
      type: object
      required: [releasedBy]
      properties:
        releasedBy:
          type: string
        notes:
          type: string

    # ── Reports ─────────────────────────────────────────────────────────────────
...