# Stock-Dashboard/lib/api-zod/src/generated/types/materialRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Warehouse Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { MaterialRequestPriority } from './materialRequestPriority';
import type { MaterialRequestStatus } from './materialRequestStatus';
import type { RequestItem } from './requestItem';

export interface MaterialRequest {
  id: number;
  requestNumber: string;
  requesterName: string;
  /** @nullable */
  requesterId?: number | null;
  department: string;
  /** @nullable */
  workLocation?: string | null;
  /** @nullable */
  jobType?: string | null;
  priority: MaterialRequestPriority;
  status: MaterialRequestStatus;
  /** @nullable */
  neededDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  approvedBy?: string | null;
  /** @nullable */
  approvalNotes?: string | null;
  items: RequestItem[];
  createdAt: string;
  updatedAt?: string;
}